+++ src/Entry/Elements/Entry.php
@@ -2254,6 +2254,7 @@ JS, [
         $parent = self::find()
             ->site('*')
             ->preferSites([$this->siteId])
+            ->unique()
             ->drafts(null)
             ->draftOf(false)
             ->status(null)
@@ -2313,6 +2314,7 @@ JS, [
                 $parentQuery = self::find()
                     ->site('*')
                     ->preferSites([$this->siteId])
+                    ->unique()
                     ->drafts(null)
                     ->draftOf(false)
                     ->status(null);