+++ src/Field/BaseRelationField.php
@@ -67,7 +67,6 @@ use GraphQL\Type\Definition\Type;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Database\Query\JoinClause;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
@@ -76,6 +75,7 @@ use Override;
 use RuntimeException;
 use Tpetry\QueryExpressions\Language\Alias;
 
+use function CraftCms\Cms\craftAuth;
 use function CraftCms\Cms\t;
 use function CraftCms\Cms\template;
 
@@ -973,17 +973,18 @@ abstract class BaseRelationField extends Field implements CrossSiteCopyableField
                 ->eagerly();
         }
 
-        $errorCount = 0;
+        $invalidTargetIds = [];
 
         foreach ($value->all() as $i => $target) {
             if (! self::_validateRelatedElement($element, $target)) {
                 /** @var Element $target */
                 $element->addModelErrors($target, "$this->handle[$i]");
-                $errorCount++;
+                $invalidTargetIds[] = $target->id;
             }
         }
 
-        if ($errorCount) {
+        if (! empty($invalidTargetIds)) {
+            $element->addInvalidNestedElementIds($invalidTargetIds);
             $selectedCount = $value->count();
             $fail(t('The selected {relatedType} {count, plural, =1{contains} other{contain}} validation errors, preventing this {type} from being saved. Edit the {relatedType} to fix them.', [
                 'relatedType' => $selectedCount === 1
@@ -1575,7 +1576,7 @@ abstract class BaseRelationField extends Field implements CrossSiteCopyableField
                 );
                 $siteIds = Arr::where($siteIds, fn ($siteId) => $siteId !== $element->siteId);
                 if (! empty($siteIds)) {
-                    $userId = Auth::id();
+                    $userId = craftAuth()->id();
                     $timestamp = now();
 
                     foreach ($siteIds as $siteId) {